+++ source/com.unity.cluster-display.sync/Runtime/QuadroSync/GfxPluginQuadroSyncSystem.cs
@@ -1,0 +1,67 @@
+﻿using System;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using UnityEngine;
+using UnityEngine.Rendering;
+
+[assembly: InternalsVisibleTo("Unity.ClusterDisplay.Sync.Editor.Tests")]
+namespace Unity.ClusterRendering
+{
+    public class GfxPluginQuadroSyncSystem
+    {
+        public enum EQuadroSyncRenderEvent
+        {
+            QuadroSyncInitialize = 0,
+            QuadroSyncQueryFrameCount,
+            QuadroSyncResetFrameCount,
+            QuadroSyncDispose,
+            QuadroSyncEnableSystem,
+            QuadroSyncEnableSwapGroup,
+            QuadroSyncEnableSwapBarrier,
+            QuadroSyncEnableSyncCounter
+        };
+
+        internal static class GfxPluginQuadroSyncUtilities
+        {
+#if (UNITY_64 || UNITY_EDITOR_64 || PLATFORM_ARCH_64)
+            [DllImport("GfxPluginQuadroSyncD3D11", CharSet = CharSet.Ansi, CallingConvention = CallingConvention.StdCall)]
+#endif
+            public static extern IntPtr GetRenderEventFuncD3D11();
+
+#if (UNITY_64 || UNITY_EDITOR_64 || PLATFORM_ARCH_64)
+            [DllImport("GfxPluginQuadroSyncD3D12", CharSet = CharSet.Ansi, CallingConvention = CallingConvention.StdCall)]
+#endif
+            public static extern IntPtr GetRenderEventFuncD3D12();
+        }
+
+        public static GfxPluginQuadroSyncSystem Instance
+        {
+            get { return instance; }
+        }
+        private static readonly GfxPluginQuadroSyncSystem instance = new GfxPluginQuadroSyncSystem();
+
+        static GfxPluginQuadroSyncSystem()
+        {
+        }
+
+        private GfxPluginQuadroSyncSystem()
+        {
+        }
+
+        public void ExecuteQuadroSyncCommand(EQuadroSyncRenderEvent id, IntPtr data)
+        {
+            var cmdBuffer = new CommandBuffer();
+
+            if (SystemInfo.graphicsDeviceType == GraphicsDeviceType.Direct3D11)
+            {
+                cmdBuffer.IssuePluginEventAndData(GfxPluginQuadroSyncUtilities.GetRenderEventFuncD3D11(), (int)id, data);
+            }
+            else if (SystemInfo.graphicsDeviceType == GraphicsDeviceType.Direct3D12)
+            {
+                cmdBuffer.IssuePluginEventAndData(GfxPluginQuadroSyncUtilities.GetRenderEventFuncD3D12(), (int)id, data);
+            }
+
+            Graphics.ExecuteCommandBuffer(cmdBuffer);
+        }
+    }
+}