+++ source/com.unity.cluster-display.sync/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changes in Cluster Display Sync
+
+## [Unreleased] - 2020-09-30
+### Changes
+- Implemented QuadroSync native plugin
+
 ## [0.0.2-preview] - 2020-03-19
 ### Changes
 - Implemented node timeouts